Move interfaces from creatures to actions
Move IAttackable and ICanattack to reduce coupling and because it makles
conceptual sense. Also remove IMovable as it is not used
