+++ src/main/java/com/group4/app/controller/PositionController.java
@@ -3,7 +3,7 @@ package com.group4.app.controller;
 import java.util.Iterator;
 
 import com.group4.app.model.Model;
-import com.group4.app.model.creatures.IAttackable;
+import com.group4.app.model.actions.IAttackable;
 import com.group4.app.model.dungeon.IPositionable;
 import com.group4.app.model.dungeon.Position;
