+++ src/main/java/com/group4/app/model/actions/AttackAction.java
@@ -4,8 +4,6 @@ import java.util.HashSet;
 import java.util.Set;
 
 import com.group4.app.model.Model;
-import com.group4.app.model.creatures.IAttackable;
-import com.group4.app.model.creatures.ICanAttack;
 import com.group4.app.model.dungeon.IPositionable;
 import com.group4.app.model.dungeon.PathfindingHelper;
 import com.group4.app.model.dungeon.Position;
