+++ src/main/java/com/group4/app/model/actions/IActionTaker.java
@@ -1,7 +1,5 @@
 package com.group4.app.model.actions;
 
-import com.group4.app.model.creatures.IAttackable;
-import com.group4.app.model.creatures.ICanAttack;
 import com.group4.app.model.dungeon.IPositionable;
 
 public interface IActionTaker extends IPositionable{
