+++ src/main/java/com/group4/app/model/actions/ICanAttack.java
@@ -1,4 +1,4 @@
-package com.group4.app.model.creatures;
+package com.group4.app.model.actions;
 
 import com.group4.app.model.dungeon.IHasPosition;
