--- src/main/java/com/group4/app/model/creatures/IMovable.java
@@ -1,4 +1,0 @@
-package com.group4.app.model.creatures;
-
-public interface IMovable {
-}
