+++ src/main/java/com/group4/app/model/input/AttackActionInput.java
@@ -1,6 +1,6 @@
 package com.group4.app.model.input;
 
-import com.group4.app.model.creatures.IAttackable;
+import com.group4.app.model.actions.IAttackable;
 
 public class AttackActionInput extends ActionInput<IAttackable> {
     public AttackActionInput(String actionId, IAttackable target) {
